requirements.txt
GitPython>=3.1.40